Rewrite fill_color() and change color data structure
We now use a bitwise shift to fill the color over the course of reading
the color string, instead of using expensive nested loops and a call to
pow().

Currently, only 24-bit colors work properly, and error-handling in the
case that a color string is not 3 or 6 characters long has been
temporarily removed.