Use SDL_WaitEvent() instead of SDL_PollEvent()
Since we only ever do something on user input, it is okay to block on
that input instead of looping over SDL_PollEvent().

The paint() helper function was added to concisely draw to the screen
only when necessary, as opposed to on every loop. Well, I realize that
right now, the loop only goes around when we need to draw to the screen,
but still. This could change and I prefer this design.