Remove MAX_SHIFT check in for loop
This check was preventing the switch statement below from working
properly. Strings greater than 6 characters would be truncated, instead
of being rejected in the switch.

Also, the CHANNELS, MAX_CHAN_CHARS, and MAX_SHIFT constants were
removed. It is essential to the program that these constants are, in
fact, 12 and 24. They are only used in the context of bit amounts once,
those usages are well commented, and it is well documented throughout
the program that 12- and 24-bit colors are supported.