Read colors in the correct order
Previously, colors would incorrectly be read in the wrong order. For
example, the color string 123456 would produce a color where r=0x21,
g=0x43, and b=65. This is now fixed.

Originally, I avoided calling strlen() to prevent us from having to read
over the string twice. However, the very small length of the
strings that we will be dealing with, and it simplifies the code by
allowing us to branch before reading the string.