Use strnlen() instead of strlen() to determine colorstring size
It doesn't matter what the size of the string is if it's greater than 6,
it's still wrong.

I've changed back to using constants here, in part to prevent the
slightly more magic "7" from being passed to strnlen().

An enum is used instead of const for the constants so that _12BITLEN and
_24BITLEN can be used as switch cases.