Add -r randomization option
This required a bit more refactoring than you might think, in part to
avoid heap allocation when generating the random color strings.

I decided to remove struct color and use SDL_Color for simplicity's
sake, but also because the order of the Uint8 fields no longer matter,
as the bit shifting section of fill_color() (renamed to strtocolor(),
with the string assignment section factored out) was unrolled. This
makes sense as we already know the length of the string.

Considerable effort was put into avoiding iterating over the string more
than necessary, and to avoid checking for the leading '#' character more
than once.

SDL initialization was factored out of main() into a separate function
init_sdl() to prevent main() from getting to long, with the added getopt
processing.